Seperate generators into seperate classes for SRP and simplified calls
