+++ .gitignore
@@ -70,3 +70,5 @@ maes-ros-slam-ws/build
 maes-ros-slam-ws/log
 # if workspace opened in pycharm, this folder will be created
 maes-ros-slam-ws/.idea
+/.vscode
+/.vsconfig
