+++ Assets/Scripts/Map/MapGen/BitMapGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f38754cf90eb9eb4fb36273c9acf88f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
