+++ Assets/Scripts/Map/MapGen/CaveGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eec839531bbf0cb4c91b134245ead083
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
