+++ Packages/manifest.json
@@ -1,13 +1,14 @@
 {
   "dependencies": {
-    "com.unity.collab-proxy": "1.15.16",
-    "com.unity.ide.rider": "3.0.14",
-    "com.unity.ide.visualstudio": "2.0.15",
+    "com.unity.ai.navigation": "1.1.5",
+    "com.unity.collab-proxy": "2.1.0",
+    "com.unity.ide.rider": "3.0.25",
+    "com.unity.ide.visualstudio": "2.0.21",
     "com.unity.ide.vscode": "1.2.5",
-    "com.unity.test-framework": "1.1.31",
+    "com.unity.test-framework": "1.1.33",
     "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.6.4",
-    "com.unity.toolchain.linux-x86_64": "1.0.0",
+    "com.unity.timeline": "1.7.5",
+    "com.unity.toolchain.linux-x86_64": "2.0.6",
     "com.unity.ugui": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
