+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.2f1
-m_EditorVersionWithRevision: 2021.3.2f1 (d6360bedb9a0)
+m_EditorVersion: 2022.3.9f1
+m_EditorVersionWithRevision: 2022.3.9f1 (ea401c316338)
